make MergeSketchWithBootloader a method recevier of arduino/builder
